delete and search works, added pics

+++ src/services/contactService.js
@@ -14,6 +14,7 @@ export const initContacts = async() => {
 			id: item.id,
 			name: item.name,
 			phone: item.phone,
+			photo: 'unavailable',
 		};
 
 		var contactString = JSON.stringify(contact);
@@ -34,25 +35,18 @@ const getOScontacts = async() => {
 		contacts = [];
 		for(var i = 0; i < data.length; i++) {
 			var contact = data[i];
-
 			var contactPhone = null;
+
 			if(contact.phoneNumbers !== undefined) {
 				contactPhone = contact.phoneNumbers[0].number;
 			}
 
-			if(contactPhone !== null) {
-				contacts.push({
-					id: contact.id,
-					name: contact.name,
-					phone: contactPhone,
-				});
-			} else {
-				contacts.push({
-					id: contact.id,
-					name: contact.name,
-					phone: '',
-				});
-			}
+			contacts.push({
+				id: contact.id,
+				name: contact.name,
+				phone: contactPhone,
+				photo: 'unavailable',
+			});
 		}
 		return contacts;
 	}
@@ -60,21 +54,34 @@ const getOScontacts = async() => {
 
 export const getAllContacts = async() => {
 	const path = FileSystem.documentDirectory;
-	var files = await FileSystem.readDirectoryAsync(path);
+	const files = await FileSystem.readDirectoryAsync(path);
 
 	let contacts = await Promise.all(files.map(async item => {
 		var contact = await FileSystem.readAsStringAsync(path + item, {
 			encoding: FileSystem.EncodingType.UTF8
 		});
+
 		contact = JSON.parse(contact)
 		return contact;
-	}))
-
+	}));
 	return contacts;
 }
 
 export async function deleteContact(id) {
-	await FileSystem.deleteAsync(contactPath,  {
-		idempotent: true
-	});
+	const path = FileSystem.documentDirectory;
+	const files = await FileSystem.readDirectoryAsync(path);
+
+	let contacts = await Promise.all(files.map(async item => {
+		const contactPath = path + item;
+		var contact = await FileSystem.readAsStringAsync(contactPath, {
+			encoding: FileSystem.EncodingType.UTF8
+		});
+
+		contact = JSON.parse(contact)
+		if(contact.id === id) {
+			await FileSystem.deleteAsync(contactPath,  {
+				idempotent: true
+			});
+		}
+	}));
 }
